requirements.txt
# --- Flask & Extensions ---
Flask>=3.0
Flask-Cors>=4.0
Flask-PyMongo>=2.3
Werkzeug>=3.1       # optional

# --- ML / AI ---
tensorflow-cpu>=2.20
numpy==1.26.4

matplotlib>=3.9
pillow>=10
h5py>=3.13
gdown>=4.7

# --- Computer Vision ---
opencv-python-headless==4.10.0.84
scikit-image>=0.24

# --- Database ---
pymongo>=4.9
dnspython>=2.6

# --- Geolocation ---
geopy>=2.4

# --- Build tools ---
pip>=25.2
setuptools
wheel
